implemented the grid
