+++ lib/dojo_web/live/book_live/book_one_live.html.heex
@@ -1,0 +1,20 @@
+<div class="flex flex-col items-center">
+  <h1 class="text-3xl font-extrabold tracking-tight text-center text-brand">Dojo Book 1</h1>
+  <p class="max-w-3xl mx-auto mt-6 text-lg text-center text-sky-600">
+    Run your friends model!
+  </p>
+  <div class="flex flex-wrap justify-around p-4">
+    <div
+      :for={{name, dis} <- @disciples |> Enum.sort_by(&elem(&1, 1).online_at, :desc)}
+      class="flex flex-col items-center justify-center justify-between inline-block w-64 h-64 mb-4 text-white transition-colors duration-200 ease-in-out border-2 shadow cursor-pointer bg-brand border-custom hover:border-red-500"
+    >
+      <%= name %>
+      <br />
+      <%= dis.node %>
+      <.live_component id="animate" name={name} module={DojoWeb.Animate} />
+    </div>
+  </div>
+  <form class="flex flex-col items-center m-0 mt-8 space-y-2" phx-change="noop" phx-submit="noop">
+    <.kernel id="3x3grid" height={224} width={224} grid_map={@grid_map} />
+  </form>
+</div>
